Areas and portals: replaced pointers with vectors and references. Fixed a couple warnings in x64.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7305 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/minizip/zip.cpp
@@ -839,7 +839,7 @@ int LoadCentralDirectoryRecord(zip64_internal* pziinit)
   {
     ZPOS64_T size_central_dir_to_read = size_central_dir;
     size_t buf_size = SIZEDATA_INDATABLOCK;
-    void* buf_read = (void*)ALLOC(buf_size);
+    void* buf_read = (void*)ALLOC((int)buf_size);
     if (ZSEEK64(pziinit->z_filefunc, pziinit->filestream, offset_central_dir + byte_before_the_zipfile, ZLIB_FILEFUNC_SEEK_SET) != 0)
       err=ZIP_ERRNO;
